+++ src/motions/ActiveMotion.h
@@ -9,18 +9,18 @@ class ActiveMotion : public Motion
 {
 public:
     virtual ~ActiveMotion();
-    uint32_t getSpeed() const;
+    float getSpeed() const;
     uint32_t getDelay() const;
     bool getAI() const;
     virtual void update(uint32_t elapsedMs);
 
 protected:
     //ActiveMotion(Element*, Motion::Type, std::istream&);
-    //ActiveMotion(Element*, Motion::Type, std::tuple<uint32_t, uint32_t, bool>);
-    ActiveMotion(Element*, Motion::Type, uint32_t speed, uint32_t delay, bool ai);
+    //ActiveMotion(Element*, Motion::Type, std::tuple<float, uint32_t, bool>);
+    ActiveMotion(Element*, Motion::Type, float speed, uint32_t delay, bool ai);
 
     uint32_t m_elapsed;
-    uint32_t m_speed;
+    float m_speed;
     uint32_t m_delay;
     bool m_ai;
 };
