+++ src/motions/AutoMotion.h
@@ -13,8 +13,8 @@ public:
     };
 
     AutoMotion(Element*, std::istream&);
-    AutoMotion(Element*, std::tuple<uint32_t, uint32_t, bool, Direction>);
-    AutoMotion(Element*, uint32_t speed, uint32_t delay, bool ai, Direction);
+    AutoMotion(Element*, std::tuple<float, uint32_t, bool, Direction>);
+    AutoMotion(Element*, float speed, uint32_t delay, bool ai, Direction);
     ~AutoMotion();
     Direction getDirection() const;
     void setElement(Element*);
